+++ UI/src/stories/molecules/RipaAgencyQuestion.stories.js
@@ -1,43 +1,20 @@
-import RipaAgencyQuestion from '@/components/molecules/RipaAgencyQuestion'
+import RipaAgencyQuestions from '@/components/molecules/RipaAgencyQuestions'
 
 export default {
-  title: 'Molecules/RipaAgencyQuestion',
-  component: RipaAgencyQuestion,
+  title: 'Molecules/RipaAgencyQuestions',
+  component: RipaAgencyQuestions,
   parameters: {},
 }
 
 export const basic = () => ({
-  components: { RipaAgencyQuestion },
+  components: { RipaAgencyQuestions },
   data() {
     return {
-      agencyQuestion: {
-        Name: 'FavoriteCar',
-        Type: 'Text',
-        Prompt: 'What is your favorite car?',
-        Hint: 'It better be a cool car too!',
-        MaxLength: 250,
-        Required: true,
+      stop: {
+        agencyQuestions: [],
       },
     }
   },
   template:
-    '<div><ripa-agency-question v-model="agencyQuestion" :question="agencyQuestion"></ripa-agency-question>{{agencyQuestion}}</div>',
-})
-
-export const notRequired = () => ({
-  components: { RipaAgencyQuestion },
-  data() {
-    return {
-      agencyQuestion: {
-        Name: 'FavoriteColor',
-        Type: 'Text',
-        Prompt: 'What is your favorite color?',
-        Hint: 'It better be a pretty color!',
-        MaxLength: 250,
-        Required: false,
-      },
-    }
-  },
-  template:
-    '<div><ripa-agency-question v-model="agencyQuestion" :question="agencyQuestion"></ripa-agency-question>{{agencyQuestion}}</div>',
+    '<div><ripa-agency-questions v-model="stop"></ripa-agency-questions>{{stop}}</div>',
 })
